introduce app insights as winston log transport (#689)
