+++ admin/package.json
@@ -69,6 +69,7 @@
     "validate": "^3.0.1",
     "validator": "^9.4.0",
     "winston": "^2.4.0",
+    "winston-azure-application-insights": "^1.3.0",
     "xregexp": "^3.2.0"
   },
   "devDependencies": {
